+++ connectlib/schemas.py
@@ -3,19 +3,42 @@
 from typing import List
 
 import numpy as np
-from pydantic import BaseModel
+import pydantic
 
 
-class ScaffoldSharedState(BaseModel):
+class _Model(pydantic.BaseModel):
+    """Base model configuration"""
+
+    class Config:
+        arbitrary_types_allowed = True
+
+
+class FedAvgAveragedState(_Model):
+    """Shared state sent by the aggregate_node in the federated
+    averaging strategy."""
+
+    avg_parameters_update: List[np.ndarray]
+
+
+class FedAvgSharedState(_Model):
+    """Shared state returned by the train method of the algorithm for each client,
+    received by the aggregate function in the federated averaging strategy.
+    """
+
+    n_samples: int
+    parameters_update: List[np.ndarray]
+
+
+class ScaffoldSharedState(_Model):
     """Shared state returned by the train method of the algorithm for each client
     (e.g. algorithms.pytorch.scaffold.train)
 
     Args:
         parameters_update (typing.List[numpy.ndarray]): the weight update of the client
             (delta between fine-tuned weights and previous weights)
-        control_variate_update (List[numpy.ndarray]): the control_variate update of the client
+        control_variate_update (typing.List[numpy.ndarray]): the control_variate update of the client
         n_samples (int): the number of samples of the client
-        server_control_variate (List[numpy.ndarray]): the server control variate (``c`` in the Scaffold paper's
+        server_control_variate (typing.List[numpy.ndarray]): the server control variate (``c`` in the Scaffold paper's
             Algo). It is sent by every client as the aggregation node doesn't have a persistent state, and
             should be the same for each client as it should not be modified in the client Algo
     """
@@ -25,11 +48,8 @@ class ScaffoldSharedState(BaseModel):
     n_samples: int
     server_control_variate: List[np.ndarray]
 
-    class Config:
-        arbitrary_types_allowed = True
 
-
-class ScaffoldAveragedStates(BaseModel):
+class ScaffoldAveragedStates(_Model):
     """Shared state sent by the aggregate_node (returned by the func strategies.scaffold.avg_shared_states)
 
     Args:
@@ -40,6 +60,3 @@ class ScaffoldAveragedStates(BaseModel):
 
     server_control_variate: List[np.ndarray]  # the new server_control_variate sent to the clients
     avg_parameters_update: List[np.ndarray]  # the weighted average of the parameters_update from each client
-
-    class Config:
-        arbitrary_types_allowed = True